Move method to file
